// src/lib/models/entities/Nutrition_InfoEntity/Nutrition_InfoEntity.ts
export type Nutrition_InfoEntity = {
  id: number;
  calories?: number;
  fat?: number;
  saturated_fat?: number;
  trans_fat?: number;
  cholesterol?: number;
  carbohydrates?: number;
  fiber?: number;
  sugars?: number;
  protein?: number;
  sodium?: number;
  vitaminD?: number;
  calcium?: number;
  iron?: number;
  potassium?: number;
};